fix: semantic versioning in release script

+++ misc/release.py
@@ -5,7 +5,10 @@ import sys
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
 from subprocess import run
-from typing import NoReturn
+from typing import Any, NoReturn
+
+
+DRY_RUN = False
 
 
 def die(msg: str) -> NoReturn:
@@ -13,8 +16,14 @@ def die(msg: str) -> NoReturn:
     sys.exit(1)
 
 
+def run_wrapper(*args: Any, **kwargs: Any) -> Any:
+    if DRY_RUN:
+        return print(f"would run: {args} {kwargs}")
+    return run(*args, **kwargs)
+
+
 def git_pull() -> None:
-    run(["git", "pull"], check=True)
+    run_wrapper(["git", "pull"], check=True)
 
 
 def check_project() -> None:
@@ -46,24 +55,36 @@ def read_new_version(current: str) -> str:
     return input("New version: ")
 
 
+def convert_version_sem(version: str) -> str:
+    if "a" in version:
+        return version.replace("a", "-alpha")
+    if "b" in version:
+        return version.replace("b", "-beta")
+    if "rc" in version:
+        return version.replace("rc", "-rc")
+    return version
+
+
 def bump_version(path: Path, old: str, new: str) -> None:
+    if DRY_RUN:
+        return print(f"would bump: {old}->{new}")
     content = path.read_text()
     path.write_text(content.replace(old, new))
 
 
 def commit_bump(path: Path, version: str) -> None:
-    run(
+    run_wrapper(
         ["git", "commit", "-m", f"chore: Bump v{version} release", str(path)],
         check=True,
     )
-    run(
+    run_wrapper(
         ["git", "tag", "-a", "-m", f"gallia v{version}", f"v{version}"],
         check=True,
     )
 
 
 def github_release(version: str) -> None:
-    run(["git", "push", "--follow-tags"], check=True)
+    run_wrapper(["git", "push", "--follow-tags"], check=True)
 
     cmd = ["gh", "release", "create", "--generate-notes"]
     pre = True if any(x in version.lower() for x in ("a", "b", "rc")) else False
@@ -71,26 +92,37 @@ def github_release(version: str) -> None:
         cmd += ["--prerelease"]
     cmd += [f"v{version}"]
 
-    run(cmd, check=True)
+    run_wrapper(cmd, check=True)
 
 
 def parse_args() -> Namespace:
     parser = ArgumentParser()
     parser.add_argument("path", type=Path, help="path to pyproject.toml")
+    parser.add_argument(
+        "-d",
+        "--dry-run_wrapper",
+        action="store_true",
+        help="dry run_wrapper, do not change anything",
+    )
     return parser.parse_args()
 
 
 def main() -> None:
     args = parse_args()
+    if args.dry_run_wrapper:
+        global DRY_RUN
+        DRY_RUN = True
+
     check_project()
     git_pull()
 
     cur_version = get_current_version(args.path)
     new_version = read_new_version(cur_version)
+    sem_new_version = convert_version_sem(new_version)
 
     bump_version(args.path, cur_version, new_version)
-    commit_bump(args.path, new_version)
-    github_release(new_version)
+    commit_bump(args.path, sem_new_version)
+    github_release(sem_new_version)
     git_pull()
